cleaned up dependencies
